sync original branch

+++ test/Permissions/Superuser/Create/AllowedTest.php
@@ -31,7 +31,9 @@ class SuperuserCreateAllowedOnlyOneTest extends GraphqlTestCase {
   public static function setUpBeforeClass()
   {
     $searchPage = Utils::pages()->get("template=search");
-    $searchPage->delete();
+    if ($searchPage->id) {
+      $searchPage->delete();
+    }
     parent::setUpBeforeClass();
   }
 
@@ -47,7 +49,7 @@ class SuperuserCreateAllowedOnlyOneTest extends GraphqlTestCase {
 
     $variables = [
       'page' => [
-        'parent' => 1,
+        'parent' => '1',
         'name' => 'search',
         'title' => 'Search'
       ]
